requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
dj-database-url==2.1.0
Django
django-cors-headers
django-filter
djangorestframework
frozenlist
gunicorn
idna==3.4
Markdown==3.4.4
multidict==6.0.4
openai==0.28.1
packaging==23.1
pdf2image==1.16.3
Pillow
PyMuPDF==1.23.4
PyMuPDFb==1.23.3
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
yarl==1.9.2
